refactor : useNavigator hook 만들기
